requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
google
langchain-google-genai
pandas

-e .